* switched to jira 4.3 and tested * cli clients improvements

+++ jira-rest-cli-runner/src/main/java/com/blogspot/leonardinius/api/impl/RhinoRegistrarImpl.java
@@ -55,6 +55,7 @@ public class RhinoRegistrarImpl implements Registrar, InitializingBean, Disposab
 
 // ------------------------ INTERFACE METHODS ------------------------
 
+
 // --------------------- Interface DisposableBean ---------------------
 
     @Override
@@ -62,7 +63,10 @@ public class RhinoRegistrarImpl implements Registrar, InitializingBean, Disposab
     {
         synchronized (lock)
         {
-            scriptService.removeEngine(engineFactory);
+            if (engineFactory != null)
+            {
+                scriptService.removeEngine(engineFactory);
+            }
             engineFactory = null;
         }
     }
